+++ app/playlist_ui/trackinfodialog.h
@@ -30,6 +30,7 @@ private:
 
   void setup_table(const Track &track);
   void setup_context_menu();
+  void setup_cover_art(const Track &track);
 
   void add_table_row(const QString& title, const QString &content);
 };